% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjustSpectra}
\alias{adjustSpectra}
\title{Adjusts a spectra, assuming that the actual spectra is not limiting}
\usage{
adjustSpectra(spectra, fractionMax, fractionMin)
}
\arguments{
\item{spectra}{a sensorSpectra object containing the spectra to be adjusted}

\item{fractionMax}{numeric [0->1] describing the preportion of limiting of
the curve corresponding to the Rmax state}

\item{fractionMin}{numeric [0->1] describing the preportion of limiting of
the curve corresponding to the Rmin state}
}
\value{
a sensorSpectra object containing the adjusted spectra
}
\description{
Adjusts a spectra, assuming that the actual spectra is not limiting
}
