% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_error_df_general}
\alias{create_error_df_general}
\title{Creates a dataframe of errors at given inaccuracies}
\usage{
create_error_df_general(
  inaccuracies,
  param_name,
  param_error_fun,
  param_min,
  param_max,
  Rmin,
  Rmax,
  delta,
  midpoint,
  ...,
  temp = 295.15,
  by = 0.01
)
}
\arguments{
\item{inaccuracies}{A vector of inaccuracies (e.g. 0.02 for 2\% error), always relative}

\item{param_name}{The name of the parameter}

\item{param_error_fun}{The function that finds the error in this parameter}

\item{param_min}{The minimum param for which to record error}

\item{param_max}{The maximum param in mV, for which to record error}

\item{Rmin}{the minimum possible ratiometric fluorescence}

\item{Rmax}{the maximum possible ratiometric fluorescence}

\item{delta}{the ratiometric fluorescence in the first wavelength}

\item{midpoint}{the sensor's midpoint}

\item{...}{Additional arguments to pass into the param_error_fun (before temperature)}

\item{temp}{(optional, default: 295.15) the temperature (in Kelvin) at which measurements were made}

\item{by}{(optional, default: 0.01) The granularity of the error table--e.g., by = 0.01 would record 275 and 275.01, etc.}
}
\value{
A dataframe of errors with columns:
param: the value of the param
'Error': the error in this param
'Inaccuracy': The inaccuracy of the measurements (relative to R).
}
\description{
Creates a dataframe of errors at given inaccuracies
}
\examples{
create_error_df_general(c(0.01, 0.02), "E", Error_E, -300, -200, 1, 5, 0.2, -275)
}
