% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_error_df_pH_multiple}
\alias{create_error_df_pH_multiple}
\title{Creates an error df at multiple inaccuracies, with multiple Rmin/Rmax/delta/pKa parameters}
\usage{
create_error_df_pH_multiple(inaccuracies, pHmin, pHmax, param_df, by = 0.01)
}
\arguments{
\item{inaccuracies}{A vector of inaccuracies (e.g. 0.02 for 2\% error), always relative}

\item{pHmin}{The minimum pH for which to record error}

\item{pHmax}{The maximum pH for which to record error}

\item{param_df}{A dataframe containing a list of sensor parameters, with these columns:
'name': An identifier for this sensor
'Rmin': the minimum possible ratiometric fluorescence for this sensor
'Rmax': the maximum possible ratiometric fluorescence for this sensor
'delta': the ratiometric fluorescence in the first wavelength for this sensor
'pKa': this sensor's midpoint/pKa}

\item{by}{(optional, default: 0.01) The granularity of the error table--e.g., by = 0.01 would record 275 and 275.01, etc.}
}
\value{
A dataframe of errors with columns:
'Name': this sensor name
'pH': the pH
'Rmin': the minimum possible ratiometric fluorescence for this sensor
'Rmax': the maximum possible ratiometric fluorescence for this sensor
'Error': the error in this pH
'Inaccuracy': The inaccuracy of the measurements (relative to R).
}
\description{
Creates an error df at multiple inaccuracies, with multiple Rmin/Rmax/delta/pKa parameters
}
\examples{
create_error_df_pH_multiple(
c(0.01, 0.02), 2, 10,
data.frame(
    "Rmin" = c(1, 2),
    "Rmax" = c(5, 6),
    "delta" = c(0.2, 1.2),
    "name" = c("normal", "plusOne"),
    "pKa" = c(7, 8)
)
)
}
