% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_error_df_pLigand}
\alias{create_error_df_pLigand}
\title{Creates a dataframe of errors in pLigand potential at given inaccuracies}
\usage{
create_error_df_pLigand(
  inaccuracies,
  pLigand_min,
  pLigand_max,
  Rmin,
  Rmax,
  delta,
  pKd,
  by = 0.01,
  ligand_name = "ligand"
)
}
\arguments{
\item{inaccuracies}{A vector of inaccuracies (e.g. 0.02 for 2\% error), always relative}

\item{pLigand_min}{The minimum pLigand for which to record error}

\item{pLigand_max}{The maximum pLigand for which to record error}

\item{Rmin}{the minimum possible ratiometric fluorescence}

\item{Rmax}{the maximum possible ratiometric fluorescence}

\item{delta}{the ratiometric fluorescence in the first wavelength}

\item{pKd}{the sensor's midpoint/pKd}

\item{by}{(optional, default: 0.01) The granularity of the error table--e.g., by = 0.01 would record 275 and 275.01, etc.}

\item{ligand_name}{The name of this ligand}
}
\value{
A dataframe of errors with columns:
'pLigand': the pLigand,
'Error': the error in this pLigand
'Inaccuracy': The inaccuracy of the measurements (relative to R).
}
\description{
Creates a dataframe of errors in pLigand potential at given inaccuracies
}
\examples{
create_error_df_pLigand(c(0.01, 0.02), 2, 10, 1, 5, 0.2, 7,
ligand_name = "NADPH")
}
