% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_ranges_multiple}
\alias{create_ranges_multiple}
\title{Takes in the input of create_error_df_redox_multiple and creates a simple ranges plot:
e.g. minimum and maximum measureable value at different error thresholds for each sensor.}
\usage{
create_ranges_multiple(
  error_df,
  thresholds = c(0.5, 1, 1.5, 2, 2.5),
  parameter = "E"
)
}
\arguments{
\item{error_df}{A dataframe of errors at least these columns:
'Name': this sensor name
'E': the redox potential (mV),
'Error': the error in this redox potential (mV)
'Inaccuracy': The inaccuracy of the measurements (relative to R).}

\item{thresholds}{A vector of error thresholds (e.g. c(0.5, 1) for 0.5mV and 1mV)}

\item{parameter}{the biochemical value being measured}
}
\value{
A dataframe of suited ranges with these columns:
'Sensor_Name': the name of the sensor
'Minimum': the minimum redox potential (mV) measurable at the given inaccuracy
'Maximum': the maximum redox potential (mV) measurable at the given inaccuracy
'Inaccuracy': the inaccuracy associated with this row (relative)
'error_thresh': the error threshold associated with this row (mV)
}
\description{
Takes in the input of create_error_df_redox_multiple and creates a simple ranges plot:
e.g. minimum and maximum measureable value at different error thresholds for each sensor.
}
\examples{
error_df <- create_error_df_redox_multiple(
  c(0.02), -400, -200,
  data.frame(
    Rmin = 0.97,
    Rmax = 4.12,
    delta = 0.23,
    name = "roGFP2",
    e0 = -299
  )
)
create_ranges_multiple(error_df)
}
