% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{E}
\alias{E}
\title{What is the redox potential (mV), given R/Rmin/Rmax/delta/midpoint/temperature?}
\usage{
E(R, Rmin, Rmax, delta, e0, temp = 295.15)
}
\arguments{
\item{R}{the ratiometric fluorescence}

\item{Rmin}{the minimum possible ratiometric fluorescence}

\item{Rmax}{the maximum possible ratiometric fluorescence}

\item{delta}{the ratiometric fluorescence in the first wavelength}

\item{e0}{the sensor's midpoint potential}

\item{temp}{the temperature at which measurements were made}
}
\value{
The redox potential (mV)
}
\description{
What is the redox potential (mV), given R/Rmin/Rmax/delta/midpoint/temperature?
}
\examples{
E(R = 3, Rmin = 1, Rmax = 5, delta = 0.2, e0 = -275)
}
