% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{error_df,redoxSensor-method}
\alias{error_df,redoxSensor-method}
\title{Finds the error df of this redox sensor at given inaccuracies}
\usage{
\S4method{error_df}{redoxSensor}(
  object,
  inaccuracies = c(0.02),
  Emin = -400,
  Emax = -200,
  temp = 295.15,
  by = 0.01,
  name = "Sensor"
)
}
\arguments{
\item{object}{A redoxSensor object}

\item{inaccuracies}{(optional, default: c(0.02)) A vector of inaccuracies
(e.g. 0.02 for 2\% error), always relative}

\item{Emin}{(optional, default: -400)  The minimum redox potential, in mV,
for which to record error}

\item{Emax}{(optional, default: -200) The maximum redox potential, in mV,
for which to record error}

\item{temp}{(optional, default: 295.15) the temperature (in Kelvin) at which measurements were made}

\item{by}{(optional, default: 0.01) The granularity of the error table--e.g.,
by = 0.01 would record 275 and 275.01, etc.}

\item{name}{(optional, default: "Sensor") A name for this sensor}
}
\value{
A dataframe of errors with columns:
'Name': this sensor name
'E': the redox potential (mV),
'Rmin': the minimum possible ratiometric fluorescence for this sensor
'Rmax': the maximum possible ratiometric fluorescence for this sensor
'Error': the error in this redox potential (mV)
'Inaccuracy': The inaccuracy of the measurements (relative to R).
}
\description{
Adding this method on 31 May 2020, hoping this style will depreciate
getErrorTable in the future.
}
\examples{
my_sensor <- new("redoxSensor", new("Sensor", Rmin = 1, Rmax = 5, delta = 0.2), e0 = -250)
error_df(my_sensor,
  inaccuracies = c(0.01, 0.02), Emin = -300, Emax = -200,
)
}
