% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{error_df}
\alias{error_df}
\title{Find the error df of an object}
\usage{
error_df(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{...}
}
\value{
A ggplot object
}
\description{
Find the error df of an object
}
