% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Error_E}
\alias{Error_E}
\title{What is the error in redox potential at a given redox potential (mV),
given some parameters of R and the error in R?}
\usage{
Error_E(E, Rmin, Rmax, delta, e0, error_R, temp = 295.15)
}
\arguments{
\item{E}{the redox potential}

\item{Rmin}{the minimum possible ratiometric fluorescence}

\item{Rmax}{the maximum possible ratiometric fluorescence}

\item{delta}{the ratiometric fluorescence in the first wavelength}

\item{e0}{the sensor's midpoint potential}

\item{error_R}{a function that, given an R, returns the error in that R}

\item{temp}{the temperature at which measurements were made}
}
\value{
A list with (1) 'E', the given redox potential,
(2) 'larger_E', the largest possible E, at the given error
(3) 'smaller_E', the smallest possible E, at the given error
(4) 'max_error' the largest possible difference between the observed and expected E
}
\description{
What is the error in redox potential at a given redox potential (mV),
given some parameters of R and the error in R?
}
\examples{
Error_E(E = -275, Rmin = 1, Rmax = 5, delta = 0.2, e0 = -275, error_R = function(x) 0.02 * x)
}
