% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Error_general}
\alias{Error_general}
\title{A general function for calculating the error in some parameter}
\usage{
Error_general(
  param,
  param_name,
  R_of_param,
  param_of_R,
  Rmin,
  Rmax,
  delta,
  midpoint,
  error_R,
  temp = 295.15
)
}
\arguments{
\item{param}{The parameter over which we are calculating error}

\item{param_name}{The name of the parameter over which we are calculating error}

\item{R_of_param}{A function that maps from param --> R
NOTE: R_of_param must have arguments in the correct ordrer. They should be:
\itemize{
\item{"param"}
\item{"Rmin"}
\item{"Rmax"}
\item{"delta"}
\item{"midpoint"}
\item{"temperature, in kelvin"}
}}

\item{param_of_R}{A function that maps from R --> param
NOTE: param_of_R must have arguments in the correct order. They should be:
#' \itemize{
\item{"R"}
\item{"Rmin"}
\item{"Rmax"}
\item{"delta"}
\item{"midpoint"}
\item{"temperature, in kelvin"}
}}

\item{Rmin}{the minimum possible ratiometric fluorescence}

\item{Rmax}{the maximum possible ratiometric fluorescence}

\item{delta}{the ratiometric fluorescence in the first wavelength}

\item{midpoint}{the sensor's midpoint (e.g. e0 for redox, pKa for pH, pKd for other ligand)}

\item{error_R}{a function that, given an R, returns the error in that R}

\item{temp}{the temperature at which measurements were made}
}
\value{
A list with (1) 'E', the given redox potential,
(2) 'larger_E', the largest possible E, at the given error
(3) 'smaller_E', the smallest possible E, at the given error
(4) 'max_error' the largest possible difference between the observed and expected E
}
\description{
A general function for calculating the error in some parameter
}
\examples{
Error_general(param = -275, param_name = "E", R_of_param = R_of_E, param_of_R = sensorOverlord::E,
Rmin = 1, Rmax = 5, delta = 0.2, midpoint = -275, error_R = function(x) 0.02 * x)
}
