% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Error_pH}
\alias{Error_pH}
\title{What is the error in pH at a given pH,
given some parameters of R and the error in R?}
\usage{
Error_pH(pH, Rmin, Rmax, delta, pKa, error_R, ...)
}
\arguments{
\item{pH}{the pH}

\item{Rmin}{the minimum possible ratiometric fluorescence}

\item{Rmax}{the maximum possible ratiometric fluorescence}

\item{delta}{the ratiometric fluorescence in the first wavelength}

\item{pKa}{the sensor's midpoint/pKa}

\item{error_R}{a function that, given an R, returns the error in that R}

\item{...}{not used (for generalization compatability)}
}
\value{
A list with (1) 'pH', the given pH,
(2) 'larger_pH', the largest possible E, at the given error
(3) 'smaller_pH', the smallest possible E, at the given error
(4) 'max_error' the largest possible difference between the observed and expected E
}
\description{
What is the error in pH at a given pH,
given some parameters of R and the error in R?
}
\examples{
Error_pH(pH = 7.3, Rmin = 1, Rmax = 5, delta = 0.2, pKa = 8, error_R = function(x) 0.02 * x)
}
