% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Error_pLigand}
\alias{Error_pLigand}
\title{What is the error in pLigand at a given pLigand,
given some parameters of R and the error in R?}
\usage{
Error_pLigand(pLigand, Rmin, Rmax, delta, pKd, error_R, ligand_name, ...)
}
\arguments{
\item{pLigand}{the pLigand}

\item{Rmin}{the minimum possible ratiometric fluorescence}

\item{Rmax}{the maximum possible ratiometric fluorescence}

\item{delta}{the ratiometric fluorescence in the first wavelength}

\item{pKd}{the sensor's midpoint/pKd}

\item{error_R}{a function that, given an R, returns the error in that R}

\item{ligand_name}{the name of the ligand}

\item{...}{not used (for generalization compatability)}
}
\value{
A list with (1) 'pH', the given pH,
(2) 'larger_pH', the largest possible E, at the given error
(3) 'smaller_pH', the smallest possible E, at the given error
(4) 'max_error' the largest possible difference between the observed and expected E
}
\description{
What is the error in pLigand at a given pLigand,
given some parameters of R and the error in R?
}
\examples{
Error_pLigand(pLigand = 7.3, ligand_name = "pNADPH",
Rmin = 1, Rmax = 5, delta = 0.2, pKd = 8, error_R = function(x) 0.02 * x)
}
