% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatSpectraData}
\alias{formatSpectraData}
\title{Formats sensor information into a dataframe suitable for input into database}
\usage{
formatSpectraData(
  name,
  type,
  readout,
  lambda_max,
  values_max,
  lambda_min,
  values_min,
  sensor_midpoint,
  lambda1_recommended,
  lambda2_recommended
)
}
\arguments{
\item{name}{Character string. The name of the sensor}

\item{type}{The type of sensor. One of:
{redox, pH, pLigand}}

\item{readout}{The readout of the sensor. One of:
{exitation ratiometric, emission ratiometric}}

\item{lambda_max}{The lambda values corresponding to the emission in the
maximum state (corresponding to Rmax)}

\item{values_max}{The emission values in the maximum state (corresponding
to Rmax--so, deprotenated or oxidized or not bound to ligand)}

\item{lambda_min}{The lambda values corresponding to the emission in the
minimum state (corresponding to Rmin)}

\item{values_min}{The emission values in the minimum state (corresponding
to Rmin--so, protenated or reduced or bound to ligand)}

\item{sensor_midpoint}{Numeric. The midpoint of the sensor. Depending
on your sensor, this could be:
{e0, pKa, log-midpoint}}
}
\value{
A list object suitable to be pushed to the mongo database
}
\description{
Formats sensor information into a dataframe suitable for input into database
}
