% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fraction_deprot}
\alias{fraction_deprot}
\title{What is the fraction deprotenated of a certain pH, given the pKa?
Note: this generalizes to any ligand-binding sensor type, given a pKa/pKd
and a pLigand}
\usage{
fraction_deprot(pH, pKa)
}
\arguments{
\item{pH}{The pH}

\item{pKa}{The pKa}
}
\description{
What is the fraction deprotenated of a certain pH, given the pKa?
Note: this generalizes to any ligand-binding sensor type, given a pKa/pKd
and a pLigand
}
