% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{getAbsError-R,redoxSensor-method}
\alias{getAbsError-R,redoxSensor-method}
\title{Get the error for a given sensor object at a given R}
\usage{
\S4method{getAbsError-R}{redoxSensor}(object, R, Error_Model, ...)
}
\arguments{
\item{object}{A sensor object}

\item{R}{A single numeric ratio value}

\item{Error_Model}{A function in the form Error_Model(R) --> Error in R}

\item{...}{Extra parameters applied to the FUN function}
}
\value{
A numeric error
}
\description{
Each R will have a certain error, defined by the Error_Model parameter
That error in R will propagate when applied to the FUN function parameter
This method returns that error
}
