% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\name{getAbsError-R}
\alias{getAbsError-R}
\title{Get the error, given a ratio value (generic)}
\usage{
`getAbsError-R`(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{...}
}
\value{
A numeric error
}
\description{
Get the error, given a ratio value (generic)
}
