% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{getAbsError,Sensor-method}
\alias{getAbsError,Sensor-method}
\title{Get the error for a given sensor object}
\usage{
\S4method{getAbsError}{Sensor}(object, R, FUN, Error_Model, ...)
}
\arguments{
\item{object}{A sensor object}

\item{R}{An array of numeric ratio values}

\item{FUN}{A function that will be applied to R}

\item{Error_Model}{A function in the form Error_Model(R) --> Error in R}

\item{...}{Extra parameters applied to the FUN function}
}
\value{
A numeric array of errors
}
\description{
Each R will have a certain error, defined by the Error_Model parameter
That error in R will propagate when applied to the FUN function parameter
This method returns that error
}
