% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{getE,redoxSensor-method}
\alias{getE,redoxSensor-method}
\title{Get the redox potential (E) for a redox sensor}
\usage{
\S4method{getE}{redoxSensor}(object, R, temp = 295.15)
}
\arguments{
\item{object}{A redoxSensor object}

\item{R}{A numeric value (can be an array) of ratio values}

\item{temp}{The temperature, in Kelvin. Default is 295.15}
}
\description{
For a given redox sensor at a certain temperature, returns the
redox potential corresponding to a given ratio (R) value
}
