% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{getE}
\alias{getE}
\title{Get the redox potential (E) value (generic)}
\usage{
getE(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{...}
}
\value{
A numeric array of E values
}
\description{
Get the redox potential (E) value (generic)
}
