% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{getE_deriv,redoxSensor-method}
\alias{getE_deriv,redoxSensor-method}
\title{Get the derivative of the redox potential (dE/dR) for a redox sensor}
\usage{
\S4method{getE_deriv}{redoxSensor}(object, R = getR(object))
}
\arguments{
\item{object}{A redoxSensor object}

\item{R}{(Optional, defaults to getR(Object)
A numeric value (can be an array) of ratio values}
}
\value{
A numeric array of dE/dR values
}
\description{
For a given redox sensor at a certain temperature, returns the
derivative of the redox potential corresponding to a given ratio (R) value
}
