% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{getE_deriv}
\alias{getE_deriv}
\title{Get the the derivative of the redox potential (dE/dR) (generic)}
\usage{
getE_deriv(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{...}
}
\value{
A numeric array of dE/dR values
}
\description{
Get the the derivative of the redox potential (dE/dR) (generic)
}
