% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{getError}
\alias{getError}
\title{Get the error (generic)}
\usage{
getError(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{...}
}
\value{
A numeric array of errors
}
\description{
Get the error (generic)
}
