% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\name{getErrorTable,Sensor-method}
\alias{getErrorTable,Sensor-method}
\title{Get the error table for a given sensor object}
\usage{
\S4method{getErrorTable}{Sensor}(object, R = getR(object), FUN = getProperty, Error_Model, ...)
}
\arguments{
\item{object}{A sensor object}

\item{R}{An array of numeric ratio values}

\item{FUN}{A function that will be applied to R}

\item{Error_Model}{A function in the form Error_Model(R) --> Error in R}

\item{...}{Extra parameters applied to the FUN function}
}
\value{
A dataframe of errors
}
\description{
Each R will have a certain error, defined by the Error_Model parameter
That error in R will propagate when applied to the FUN function parameter
This method returns (1) The input Rs (2) the error in Rs
(3) the results of the FUN function
parameter (4) the upper bound of the FUN function (5) the lower bound
of te fun function (6) the max absolute error for the FUN function
}
