% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{getErrorTable}
\alias{getErrorTable}
\title{Get a table of the errors (generic)}
\usage{
getErrorTable(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{...}
}
\value{
A dataframe of errors
}
\description{
Get a table of the errors (generic)
}
