% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{getFractionMax,Sensor-method}
\alias{getFractionMax,Sensor-method}
\title{Get the fraction of sensors in the state corresponding to "Rmax"}
\usage{
\S4method{getFractionMax}{Sensor}(object, R)
}
\arguments{
\item{object}{A sensor object}

\item{R}{A single number or numeric vector corresponding to an R between Rmin and Rmax}

\item{...}{...}
}
\value{
A single number of numeric vector of the fraction of sensors
in the maximum-emission state corresponding to the given R
}
\description{
This method corresponds to finding the "OxD" in a redox sensor
}
\examples{
my_sensor <- new("Sensor", Rmin = 1, Rmax = 5, delta = 0.5)
getFractionMax(my_sensor, R = 1)
getFractionMax(my_sensor, R = 3)
getFractionMax(my_sensor, R = 5)
}
