% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\name{getFractionMax}
\alias{getFractionMax}
\title{Generic for the getFractionMax method}
\usage{
getFractionMax(object, ...)
}
\arguments{
\item{object}{A sensor object}

\item{...}{...}
}
\value{
A generic function for getFractionMax
}
\description{
This method corresponds to finding the "OxD" in a redox sensor
}
