% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{getpH}
\alias{getpH}
\title{Get the pH value (generic)}
\usage{
getpH(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{...}
}
\value{
A numeric array of pH values
}
\description{
Get the pH value (generic)
}
