% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{getProperty,Sensor-method}
\alias{getProperty,Sensor-method}
\title{Get the fraction of sensors in the state corresponding to "Rmax" (wrapper)}
\usage{
\S4method{getProperty}{Sensor}(object, ...)
}
\arguments{
\item{object}{A sensor object}

\item{...}{...}
}
\value{
A single number of numeric vector of the fraction of sensors
in the maximum-emission state corresponding to the given R
}
\description{
Get the fraction of sensors in the state corresponding to "Rmax" (wrapper)
}
\examples{
my_sensor <- new("Sensor", Rmin = 1, Rmax = 5, delta = 0.5)
getFractionMax(my_sensor, R = 1)
getFractionMax(my_sensor, R = 3)
getFractionMax(my_sensor, R = 5)
}
