% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\name{getProperty,pHSensor-method}
\alias{getProperty,pHSensor-method}
\title{Get the pH of a pH sensor (wrapper)}
\usage{
\S4method{getProperty}{pHSensor}(object, ...)
}
\arguments{
\item{object}{A pHSensor object}

\item{...}{...}
}
\value{
A numeric array of pH values
}
\description{
For a given redox sensor at a certain temperature, returns the
redox potential corresponding to a given ratio (R) value
}
