% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{getProperty,pHSensor-method}
\alias{getProperty,pHSensor-method}
\title{Get the pH of a pH sensor}
\usage{
\S4method{getProperty}{pHSensor}(object, R = getR(object))
}
\arguments{
\item{object}{A pHSensor object}

\item{R}{(Optional, defaults to getR(Object)
A numeric value (can be an array) of ratio values}
}
\value{
A numeric array of pH values
}
\description{
For a given redox sensor at a certain temperature, returns the
redox potential corresponding to a given ratio (R) value
}
