% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{getProperty,redoxSensor-method}
\alias{getProperty,redoxSensor-method}
\title{Get the redox potential (E) for a redox sensor (wrapper)}
\usage{
\S4method{getProperty}{redoxSensor}(object, ...)
}
\arguments{
\item{object}{A redoxSensor object}

\item{...}{...}
}
\value{
A numeric array of E values
}
\description{
For a given redox sensor at a certain temperature, returns the
redox potential corresponding to a given ratio (R) value
}
