% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{getProperty}
\alias{getProperty}
\title{Wrapper: Get the function that gets the biochemical property of this sensor}
\usage{
getProperty(object, ...)
}
\arguments{
\item{object}{A sensor-type object}

\item{...}{...}
}
\value{
A function that gets the biochemical property of this sensor
}
\description{
Wrapper: Get the function that gets the biochemical property of this sensor
}
