% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{getR,Sensor-method}
\alias{getR,Sensor-method}
\title{Method to get an array of R values from a sensor}
\usage{
\S4method{getR}{Sensor}(object, by = 0.01)
}
\arguments{
\item{object}{A sensor object}

\item{by}{The "by" argument for seq. Specifies intervals between subsequent R values}
}
\value{
A numeric array between Rmin and Rmax
}
\description{
Takes in a sensor object and returns a vector of valid R values
}
\examples{
my_sensor <- new("Sensor", Rmin = 1, Rmax = 5, delta = 0.5)
getR(my_sensor, by = 0.1)
}
