% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{getR}
\alias{getR}
\title{Get an array of R values from a sensor}
\usage{
getR(object, ...)
}
\arguments{
\item{object}{A sensor object}

\item{...}{...}

\item{by}{The "by" argument for seq. Specifies intervals between subsequent R values}
}
\value{
A numeric array between Rmin and Rmax
}
\description{
This is just a simple method that generates a range of values between Rmin and Rmax
}
\examples{
my_sensor <- new("Sensor", Rmin = 1, Rmax = 5, delta = 0.5)
getR(my_sensor, by = 0.1)
}
