% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{getR}
\alias{getR}
\title{Get an array of R values (generic)}
\usage{
getR(object, ...)
}
\arguments{
\item{object}{A sensor object}

\item{...}{...}
}
\value{
A numeric array
}
\description{
This is just a simple method that generates a
range of values between Rmin and Rmax
}
