% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Classes.R
\docType{class}
\name{ligandSensor-class}
\alias{ligandSensor-class}
\title{An S4 class to represent a 2-state ligand sensor}
\description{
An S4 class to represent a 2-state ligand sensor
}
\section{Slots}{

\describe{
\item{\code{Rmin}}{To represent the ratio emission value R in the bound state}

\item{\code{Rmax}}{to represent the ratio emission value R in the unbound state}

\item{\code{delta}}{To represent the ratio between emission in the
unbound and bound states in the second wavelength of the ratio.}

\item{\code{pKd}}{The midpoint/pKd of the ligand sensor}
}}

