% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Constructors.R
\name{newSensorFromSpectra}
\alias{newSensorFromSpectra}
\title{A function to convert a sensorSpectra to a Sensor}
\usage{
newSensorFromSpectra(sensorSpectra, lambda_1, lambda_2)
}
\arguments{
\item{sensorSpectra}{A sensorSpectra object}

\item{lambda_1}{the first wavelength in the
excitation ratio lambda_1/lambda_2}

\item{lambda_2}{the second wavelength in the
excitation ratio lambda_1/lambda_2}
}
\value{
A Sensor object
}
\description{
Converts a sensorSpectra object to a Sensor object.
A sensor object exists alongside a wavelength ratio (e.g. 410/470),
so this function must be supplied those (in the form lambda_1/lambda_2)
}
