% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pH}
\alias{pH}
\title{Finds pH, given R, Rmin, and Rmax}
\usage{
pH(R, Rmin, Rmax, delta, pKa, ...)
}
\arguments{
\item{R}{the ratiometric fluorescence}

\item{Rmin}{the minimum possible ratiometric fluorescence}

\item{Rmax}{the maximum possible ratiometric fluorescence}

\item{delta}{the ratiometric fluorescence in the first wavelength}

\item{pKa}{the sensor's pKa}

\item{...}{unused}
}
\value{
the pH, numeric
}
\description{
Finds pH, given R, Rmin, and Rmax
}
\examples{
pH(R = 2, Rmin = 1, Rmax = 5, delta = 0.2, pKa = 7)
}
