% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Classes.R
\docType{class}
\name{pHSensor-class}
\alias{pHSensor-class}
\title{An S4 class to represent a 2-state pH sensor}
\description{
An S4 class to represent a 2-state pH sensor
}
\section{Slots}{

\describe{
\item{\code{Rmin}}{To represent the ratio emission value R in the protenated state}

\item{\code{Rmax}}{to represent the ratio emission value R in the deprotenated state}

\item{\code{delta}}{To represent the ratio between emission in the
deprotenated and protenated states in the second wavelength of the ratio.}

\item{\code{pKa}}{The midpoint/pKa of the pH sensor}
}}

