% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pLigand}
\alias{pLigand}
\title{Finds pLigand, given R, Rmin, and Rmax}
\usage{
pLigand(R, Rmin, Rmax, delta, pKd, ...)
}
\arguments{
\item{R}{the ratiometric fluorescence}

\item{Rmin}{the minimum possible ratiometric fluorescence}

\item{Rmax}{the maximum possible ratiometric fluorescence}

\item{delta}{the ratiometric fluorescence in the first wavelength}

\item{pKd}{the sensor's pKd}

\item{...}{not used}
}
\value{
the pLigand, numeric
}
\description{
Finds pLigand, given R, Rmin, and Rmax
}
\examples{
pLigand(R = 2, Rmin = 1, Rmax = 5, delta = 0.2, pKd = 7)
}
