% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_ranges_general}
\alias{plot_ranges_general}
\title{Takes in a ranges_df dataframe and makes a plot!}
\usage{
plot_ranges_general(ranges, ylim, by, y_label)
}
\arguments{
\item{ranges}{A dataframe of ranges with at least these columns:
'Sensor_Name': the name of the sensor
'Minimum': the minimum parameter measurable at the given inaccuracy
'Maximum': the maximum parameter measurable at the given inaccuracy
'Inaccuracy': the inaccuracy associated with this row (relative)
'error_thresh': the error threshold associated with this row}

\item{ylim}{The limits of the ranges plot}

\item{by}{the 'by' argument of the limits axis tick marks}
}
\value{
A ggplot object
}
\description{
Takes in a ranges_df dataframe and makes a plot!
}
