% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_ranges_pLigand}
\alias{plot_ranges_pLigand}
\title{Takes in a ranges_df dataframe and makes a plot (for pLigand).}
\usage{
plot_ranges_pLigand(ranges, ylim = c(1, 14), by = 1, ylab = "pLigand")
}
\arguments{
\item{ranges}{A dataframe of ranges with at least these columns:
'Sensor_Name': the name of the sensor
'Minimum': the minimum pLigand measurable at the given inaccuracy
'Maximum': the maximum pLigand measurable at the given inaccuracy
'Inaccuracy': the inaccuracy associated with this row (relative)
'error_thresh': the error threshold associated with this row}

\item{ylim}{The limits of the ranges plot}

\item{by}{the 'by' argument of the limits axis tick marks}

\item{ylab}{The label of the ranges plot}
}
\value{
A ggplot object
}
\description{
Takes in a ranges_df dataframe and makes a plot (for pLigand).
}
\examples{
error_df <- create_error_df_pLigand_multiple(
c(0.01, 0.02), 2, 10,
data.frame(
  "Rmin" = c(1, 2),
  "Rmax" = c(5, 6),
  "delta" = c(0.2, 1.2),
  "name" = c("normal", "plusOne"),
  "pKd" = c(7, 8)
),
ligand_name = "NADPH"
)
ranges_df <- create_ranges_multiple(error_df, parameter = "NADPH",
thresholds = c(0.01, 0.05, 0.10, 0.15, 0.20))
plot_ranges_pLigand(ranges_df, ylab = "pNADPH")
}
