% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{plotE,redoxSensor-method}
\alias{plotE,redoxSensor-method}
\title{Plot the fraction of E of a redoxSensor}
\usage{
\S4method{plotE}{redoxSensor}(object)
}
\arguments{
\item{object}{An redoxSensor object}
}
\value{
A ggplot object
}
\description{
Creates a ggplot object that has the ratio R on the horizontal axis
and the redox potential (E) on the vertical axis
}
