% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{plotE_deriv}
\alias{plotE_deriv}
\title{Plot the fraction of derivative of E of an object}
\usage{
plotE_deriv(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{...}
}
\value{
A ggplot object
}
\description{
Plot the fraction of derivative of E of an object
}
