% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\name{plotFractionMax,pHSensor-method}
\alias{plotFractionMax,pHSensor-method}
\title{Plot the fraction of pH sensors in the max state}
\usage{
\S4method{plotFractionMax}{pHSensor}(object, R = getR(object))
}
\arguments{
\item{object}{An pHSensor object}

\item{R}{An array of numeric ratio values}
}
\value{
A ggplot object
}
\description{
Creates a ggplot object that has the ratio R on the horizontal axis
and the fraction deprotenated on the vertical axis
}
