% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{plotFractionMax,redoxSensor-method}
\alias{plotFractionMax,redoxSensor-method}
\title{Plot the fraction of redox sensors in the max state}
\usage{
\S4method{plotFractionMax}{redoxSensor}(object)
}
\arguments{
\item{object}{An redoxSensor object}
}
\value{
A ggplot object
}
\description{
Creates a ggplot object that has the ratio R on the horizontal axis
and the fraction oxidizied on the vertical axis
}
