% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{plotFractionMax,Sensor-method}
\alias{plotFractionMax,Sensor-method}
\title{Plot the fraction of sensors in the max state}
\usage{
\S4method{plotFractionMax}{Sensor}(object, FUN)
}
\arguments{
\item{object}{An sensor object}

\item{FUN}{A function in the form FUN(Sensor, R) --> Fraction in Max State}
}
\value{
A ggplot object
}
\description{
Creates a ggplot object that has the ratio R on the horizontal axis
and the fraction of sensors corresponding to the Rmax state on the vertical axis
}
