% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{plotFractionMax}
\alias{plotFractionMax}
\title{Plot the fraction of objects in the max state (generic)}
\usage{
plotFractionMax(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{...}
}
\value{
A ggplot object
}
\description{
Plot the fraction of objects in the max state (generic)
}
