% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\name{plotProperty,pHSensor-method}
\alias{plotProperty,pHSensor-method}
\title{Plot the fraction of pH of a pHSensor}
\usage{
\S4method{plotProperty}{pHSensor}(object, R = getR(object))
}
\arguments{
\item{object}{An pHSensor object}

\item{R}{An array of numeric ratio values}
}
\value{
A ggplot object
}
\description{
Creates a ggplot object that has the ratio R on the horizontal axis
and the pH on the vertical axis
}
