% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\name{plotProperty,Sensor-method}
\alias{plotProperty,Sensor-method}
\title{Plot the fraction max of a generic}
\usage{
\S4method{plotProperty}{Sensor}(object, ...)
}
\arguments{
\item{object}{An Sensor object}

\item{...}{...}
}
\value{
A ggplot object
}
\description{
Plot the fraction max of a generic
}
