% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{plotProperty,Sensor-method}
\alias{plotProperty,Sensor-method}
\title{Plot the fraction max of a generic}
\usage{
\S4method{plotProperty}{Sensor}(object, R = getR(object))
}
\arguments{
\item{object}{An Sensor object}

\item{R}{An array of numeric ratio values}
}
\value{
A ggplot object
}
\description{
Plot the fraction max of a generic
}
