% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{plotProperty}
\alias{plotProperty}
\title{Plot the property of an object}
\usage{
plotProperty(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{...}
}
\value{
A ggplot object
}
\description{
Plot the property of an object
}
