% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Methods.R
\docType{methods}
\name{plotSpectra}
\alias{plotSpectra}
\title{A generic for the plotSpectra method}
\usage{
plotSpectra(object, ...)
}
\arguments{
\item{object}{A spectra object}

\item{...}{...}
}
\value{
A ggplot object
}
\description{
A generic for the plotSpectra method
}
