% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Methods.R
\docType{methods}
\name{plotSpectra}
\alias{plotSpectra}
\title{Create a spectra plot}
\usage{
plotSpectra(object, ...)
}
\arguments{
\item{object}{A spectra object}

\item{...}{...}
}
\value{
A ggplot object
}
\description{
This just generates a ggplot object of the emission spectra from a sensorSpectra object
}
