% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Methods.R
\docType{methods}
\name{plotSpectra,sensorSpectra-method}
\alias{plotSpectra,sensorSpectra-method}
\title{Create a plot (ggplot object) for a sensorSpetra object}
\usage{
\S4method{plotSpectra}{sensorSpectra}(object,
  minimum_name = "Minimum State", maximum_name = "Maximum State")
}
\arguments{
\item{object}{A sensorSpectra object}

\item{minimum_name}{The legend label for the state corresponding to Rmin}

\item{maximum_name}{The legend label for the state correspondign to Rmax}
}
\value{
A ggplot object
}
\description{
Returns a ggplot object plotting the emission (relative)
as a function of wavelength (lambda) for a given sensorSpectra object
}
