% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{R_of_pH}
\alias{R_of_pH}
\title{Finds R, given a pH}
\usage{
R_of_pH(pH, Rmin, Rmax, delta, pKa)
}
\arguments{
\item{pH}{the sensor's pH}

\item{Rmin}{the minimum possible ratiometric fluorescence}

\item{Rmax}{the maximum possible ratiometric fluorescence}

\item{delta}{the ratiometric fluorescence in the first wavelength}

\item{pKa}{the sensor's pKa}
}
\value{
the ratiometric fluoresence at this pH
}
\description{
Finds R, given a pH
}
\examples{
R_of_pH(pH = 3, Rmin = 1, Rmax = 5, delta = 0.2, pKa = 7)
}
