% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{R_of_pLigand}
\alias{R_of_pLigand}
\title{Finds R, given a pLigand}
\usage{
R_of_pLigand(pLigand, Rmin, Rmax, delta, pKd, ...)
}
\arguments{
\item{pLigand}{the sensor's pLigand}

\item{Rmin}{the minimum possible ratiometric fluorescence}

\item{Rmax}{the maximum possible ratiometric fluorescence}

\item{delta}{the ratiometric fluorescence in the first wavelength}

\item{pKd}{the sensor's pKd}
}
\value{
the ratiometric fluoresence at this pH
}
\description{
Finds R, given a pLigand
}
\examples{
R_of_pLigand(pLigand = 2, Rmin = 1, Rmax = 5, delta = 0.2, pKd = 7)
}
