% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\name{rangePlot,ligandSensor-method}
\alias{rangePlot,ligandSensor-method}
\title{Make a plot of the suited ranges for this ligandSensor}
\usage{
\S4method{rangePlot}{ligandSensor}(object, ranges = ranges_df(object), ylim = c(1, 14), by = 1)
}
\arguments{
\item{object}{A ligandSensor object}

\item{ranges}{(optional, default = ranges_df(object)) A ranges dataframe)}

\item{ylim}{The limits of the ranges plot}

\item{by}{the 'by' argument of the limits axis tick marks}
}
\value{
A dataframe of suited ranges with these columns:
'Sensor_Name': the name of the sensor
'Minimum': the minimum pLigand measurable at the given inaccuracy
'Maximum': the maximum pLigand measurable at the given inaccuracy
'Inaccuracy': the inaccuracy associated with this row (relative)
'error_thresh': the error threshold associated with this row (mV)
}
\description{
Make a plot of the suited ranges for this ligandSensor
}
\examples{
my_sensor <- new("ligandSensor", new("Sensor", Rmin = 1, Rmax = 5, delta = 0.2),
pKd = 7, ligand_name = "NADPH")
rangePlot(my_sensor)
}
