% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\name{rangePlot,pHSensor-method}
\alias{rangePlot,pHSensor-method}
\title{Make a plot of the suited ranges for this pHSensor}
\usage{
\S4method{rangePlot}{pHSensor}(object, ranges = ranges_df(object), ylim = c(1, 14), by = 1)
}
\arguments{
\item{object}{A pHSensor object}

\item{ranges}{(optional, default = ranges_df(object)) A ranges dataframe)}

\item{ylim}{The limits of the ranges plot}

\item{by}{the 'by' argument of the limits axis tick marks}
}
\value{
A dataframe of suited ranges with these columns:
'Sensor_Name': the name of the sensor
'Minimum': the minimum pH measurable at the given inaccuracy
'Maximum': the maximum pH measurable at the given inaccuracy
'Inaccuracy': the inaccuracy associated with this row (relative)
'error_thresh': the error threshold associated with this row (mV)
}
\description{
Make a plot of the suited ranges for this pHSensor
}
\examples{
my_sensor <- new("pHSensor", new("Sensor", Rmin = 1, Rmax = 5, delta = 0.2), pKa = 7)
rangePlot(my_sensor)
}
