% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{rangePlot,redoxSensor-method}
\alias{rangePlot,redoxSensor-method}
\title{Make a plot of the suited ranges for this sensor}
\usage{
\S4method{rangePlot}{redoxSensor}(object, ranges = ranges_df(object), ylim = c(-350, -150), by = 20)
}
\arguments{
\item{object}{A redoxSensor object}

\item{ranges}{(optional, default = ranges_df(object)) A ranges dataframe)}

\item{ylim}{The limits of the ranges plot}

\item{by}{the 'by' argument of the limits axis tick marks}
}
\value{
A dataframe of suited ranges with these columns:
'Sensor_Name': the name of the sensor
'Minimum': the minimum redox potential (mV) measurable at the given inaccuracy
'Maximum': the maximum redox potential (mV) measurable at the given inaccuracy
'Inaccuracy': the inaccuracy associated with this row (relative)
'error_thresh': the error threshold associated with this row (mV)
}
\description{
Make a plot of the suited ranges for this sensor
}
\examples{
my_sensor <- new("redoxSensor", new("Sensor", Rmin = 1, Rmax = 5, delta = 0.2), e0 = -250)
rangePlot(my_sensor)
}
