% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{rangePlot}
\alias{rangePlot}
\title{Make a rangeplot for this object}
\usage{
rangePlot(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{...}
}
\value{
A ggplot object
}
\description{
Make a rangeplot for this object
}
