% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\name{ranges_df,redoxSensor-method}
\alias{ranges_df,redoxSensor-method}
\title{Finds the ranges df of this redox sensor at given inaccuracies}
\usage{
\S4method{ranges_df}{redoxSensor}(
  object,
  inaccuracies = c(0.02),
  Emin = -400,
  Emax = -200,
  temp = 295.15,
  by = 0.01,
  name = "Sensor",
  thresholds = c(0.5, 1, 1.5, 2, 2.5)
)
}
\arguments{
\item{object}{A redoxSensor object}

\item{inaccuracies}{(optional, default: c(0.02)) A vector of inaccuracies
(e.g. 0.02 for 2\% error), always relative}

\item{Emin}{(optional, default: -400)  The minimum redox potential, in mV,
for which to record error}

\item{Emax}{(optional, default: -200) The maximum redox potential, in mV,
for which to record error}

\item{temp}{(optional, default: 295.15) the temperature (in Kelvin) at which measurements were made}

\item{by}{(optional, default: 0.01) The granularity of the error table--e.g.,
by = 0.01 would record 275 and 275.01, etc.}

\item{name}{(optional, default: "Sensor") A name for this sensor}

\item{thresholds}{A vector of error thresholds (e.g. c(0.5, 1) for 0.5mV and 1mV)}
}
\value{
A dataframe of suited ranges with these columns:
'Sensor_Name': the name of the sensor
'Minimum': the minimum redox potential (mV) measurable at the given inaccuracy
'Maximum': the maximum redox potential (mV) measurable at the given inaccuracy
'Inaccuracy': the inaccuracy associated with this row (relative)
'error_thresh': the error threshold associated with this row (mV)
}
\description{
Adding this method on 31 May 2020, hoping this style will depreciate
getErrorTable in the future.
}
\examples{
my_sensor <- new("redoxSensor", new("Sensor", Rmin = 1, Rmax = 5, delta = 0.2), e0 = -250)
ranges_df(my_sensor)
}
