% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\name{ranges_df}
\alias{ranges_df}
\title{Find the ranges df of an object}
\usage{
ranges_df(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{...}
}
\value{
A ranges dataframe
}
\description{
Find the ranges df of an object
}
