% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Classes.R
\docType{class}
\name{redoxSensor-class}
\alias{redoxSensor-class}
\title{An S4 class to represent a 2-state redox sensor}
\description{
An S4 class to represent a 2-state redox sensor
}
\section{Slots}{

\describe{
\item{\code{e0}}{The midpoint potential of the redox sensor}
}}

\examples{
testSensor <- new("redoxSensor",
                    new("Sensor", Rmin = 1, Rmax = 5, delta = 0.2), e0 = -275)
}
