% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescaleToRange}
\alias{rescaleToRange}
\title{A function to rescale a certain (x,y) matrix pair into a new (x_new, y) pair.}
\usage{
rescaleToRange(new_xs, old_xs, y)
}
\arguments{
\item{new_xs}{Numeric vector. The new x labels that we want to rescale to}

\item{old_xs}{Numeric vector. The x labels of the previous matrix}

\item{y}{Numeric vector. The original y values corresponding to old_xs}
}
\value{
Numeric vector corresponding to the y's rescaled to fit the x_new vector
}
\description{
A function to rescale a certain (x,y) matrix pair into a new (x_new, y) pair.
}
