% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Classes.R
\docType{class}
\name{Sensor-class}
\alias{Sensor-class}
\title{An S4 class to represent a 2-state sensor}
\description{
An S4 class to represent a 2-state sensor
}
\section{Slots}{

\describe{
\item{\code{Rmin}}{To represent the ratio emission value R in the minimum state}

\item{\code{Rmax}}{to represent the ratio emission value R in the maximum state}

\item{\code{delta}}{To represent the ratio between emission in the maximum and
minimum states in the second wavelength of the ratio.}
}}

