% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSensorOverlordApp.R
\name{sensorOverlordApp}
\alias{sensorOverlordApp}
\title{A wrapper function to run the included Shiny App}
\usage{
sensorOverlordApp()
}
\description{
A wrapper function to run the included Shiny App
}
