% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Class.R
\docType{class}
\name{sensorSpectra-class}
\alias{sensorSpectra-class}
\title{An S4 class to represent the emission spectrum of a 2-state sensor}
\description{
An S4 class to represent the emission spectrum of a 2-state sensor
}
\section{Slots}{

\describe{
\item{\code{lambdas}}{represents excitation wavelengths in nanometers}

\item{\code{values_minimum}}{represents the relative emission values of the sensor state
corresponding to the lowest R value Rmin}

\item{\code{values_maximum}}{represents the relative emission values of the sensor state
corresponding to the highest R value Rmax}
}}

