% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Constructors.R
\name{spectraMatrixFromValues}
\alias{spectraMatrixFromValues}
\title{A function to make a sensorSpectra from a set of 4 vectors}
\usage{
spectraMatrixFromValues(lambdas_minimum, values_minimum, lambdas_maximum,
  values_maximum)
}
\arguments{
\item{lambdas_minimum}{Lambdas of the values corresponding to the Rmin state}

\item{values_minimum}{Emission values corresponding to the Rmin state}

\item{lambdas_maximum}{Lambdas of the values corresponding to the Rmax state}

\item{values_maximum}{Emission values corresponding to the Rmax state}
}
\value{
A sensorSpectra object
}
\description{
I'll typically collect spectra data in an excel file with 4 columns: the lambda and emission
values corresponding to the minimum and maximum states. This function takes data from those 4
columns, condenses them into a 3-column dataframe (with matching lambda values) and makes
a sensorSpectra object
}
