% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_ranges_multiple}
\alias{create_ranges_multiple}
\title{Takes in the input of create_error_df_redox_multiple and creates a simple ranges plot:
e.g. minimum and maximum measureable value at different error thresholds for each sensor.}
\usage{
create_ranges_multiple(error_df, thresholds = c(0.5, 1, 1.5, 2, 2.5))
}
\arguments{
\item{error_df}{A dataframe of errors at least these columns:
'Name': this sensor name
'E': the redox potential (mV),
'Error': the error in this redox potential (mV)
'Inaccuracy': The inaccuracy of the measurements (relative to R).}

\item{thresholds}{A vector of error thresholds (e.g. c(0.5, 1) for 0.5mV and 1mV)}
}
\description{
Takes in the input of create_error_df_redox_multiple and creates a simple ranges plot:
e.g. minimum and maximum measureable value at different error thresholds for each sensor.
}
