% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{getProperty,redoxSensor-method}
\alias{getProperty,redoxSensor-method}
\title{Get the redox potential (E) for a redox sensor}
\usage{
\S4method{getProperty}{redoxSensor}(object, R = getR(object),
  temp = 295.15)
}
\arguments{
\item{object}{A redoxSensor object}

\item{R}{(Optional, defaults to getR(Object)
A numeric value (can be an array) of ratio values}

\item{temp}{The temperature, in Kelvin. Default is 295.15}
}
\value{
A numeric array of E values
}
\description{
For a given redox sensor at a certain temperature, returns the
redox potential corresponding to a given ratio (R) value
}
