% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Methods.R
\docType{methods}
\name{plotE_deriv,redoxSensor-method}
\alias{plotE_deriv,redoxSensor-method}
\title{Plot the derivative of E of a redoxSensor}
\usage{
\S4method{plotE_deriv}{redoxSensor}(object)
}
\arguments{
\item{object}{An redoxSensor object}
}
\value{
A ggplot object
}
\description{
Creates a ggplot object that has the ratio R on the horizontal axis
and the derivative of redox potential (dE/dR) on the vertical axis
}
